Synapse_Agent/requirements.txt
# // filepath: d:\Project Based Learning\Synapse_Agent\requirements.txt
langchain
wikipedia
langchain-community
google-generativeai
langchain-google-genai
python-dotenv 
pydantic